do not precompute pixels

+++ src/osrs/tables/texture_pixels.h
@@ -10,7 +10,7 @@ int* texture_pixels_new_from_definition(
     struct SpritePack* sprite_packs,
     int* pack_ids,
     int pack_count,
-    int brightness);
+    double brightness);
 void texture_pixels_free(struct TexturePixels* pixels);
 
 #endif
